# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для мобильных складских сценариев.",
        "title": "API сервиса склада",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/boxes/move": {
            "post": {
                "description": "Перемещает короб по коду маркера в целевую ячейку хранения и записывает операцию в историю.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Короба"
                ],
                "summary": "Переместить короб в ячейку",
                "parameters": [
                    {
                        "description": "Данные для перемещения короба",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MoveBoxRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Короб успешно перемещён",
                        "schema": {
                            "$ref": "#/definitions/models.MoveBoxResult"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Объект не найден",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Конфликт состояния",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/objects/by-marker": {
            "get": {
                "description": "Возвращает унифицированную карточку объекта для указанного маркера.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Объекты"
                ],
                "summary": "Получить объект по коду маркера",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Код маркера",
                        "name": "marker_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Карточка объекта",
                        "schema": {
                            "$ref": "#/definitions/models.ObjectCard"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Объект не найден",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/operations": {
            "get": {
                "description": "Возвращает последние складские операции.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Операции"
                ],
                "summary": "Получить историю операций",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Максимальное число записей; по умолчанию 50, максимум 200",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список операций",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.OperationHistory"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Сохраняет информацию об операции, выполненной над складским объектом.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Операции"
                ],
                "summary": "Создать запись истории операций",
                "parameters": [
                    {
                        "description": "Данные операции",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Операция сохранена",
                        "schema": {
                            "$ref": "#/definitions/models.OperationHistory"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scan": {
            "post": {
                "description": "Находит объект по коду маркера и записывает событие сканирования.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Сканирование"
                ],
                "summary": "Сканировать объект по маркеру",
                "parameters": [
                    {
                        "description": "Данные сканирования",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ScanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат сканирования",
                        "schema": {
                            "$ref": "#/definitions/models.ScanResult"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Объект не найден",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scan-events": {
            "get": {
                "description": "Возвращает последние события сканирования маркеров.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "События сканирования"
                ],
                "summary": "Получить события сканирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Максимальное число записей; по умолчанию 50, максимум 200",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список событий сканирования",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ScanEvent"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Сохраняет попытку сканирования по коду маркера.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "События сканирования"
                ],
                "summary": "Создать событие сканирования",
                "parameters": [
                    {
                        "description": "Данные события сканирования",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateScanEventRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Событие сканирования сохранено",
                        "schema": {
                            "$ref": "#/definitions/models.ScanEvent"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "description": "Проверяет готовность API через ping пула подключений PostgreSQL.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Система"
                ],
                "summary": "Проверка доступности",
                "responses": {
                    "200": {
                        "description": "Сервис доступен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "База данных недоступна",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.CreateOperationRequest": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object"
                },
                "object_id": {
                    "type": "integer"
                },
                "object_type": {
                    "type": "string"
                },
                "operation_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "handler.CreateScanEventRequest": {
            "type": "object",
            "properties": {
                "device_info": {
                    "type": "string"
                },
                "marker_code": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
        "handler.MoveBoxRequest": {
            "type": "object",
            "properties": {
                "box_marker_code": {
                    "type": "string"
                },
                "to_storage_cell_marker_code": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "handler.ScanRequest": {
            "type": "object",
            "properties": {
                "device_info": {
                    "type": "string"
                },
                "marker_code": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.MoveBoxResult": {
            "type": "object",
            "properties": {
                "box": {
                    "$ref": "#/definitions/models.ObjectCard"
                },
                "operation": {
                    "$ref": "#/definitions/models.OperationHistory"
                }
            }
        },
        "models.ObjectCard": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "location_code": {
                    "type": "string"
                },
                "marker_code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "object_id": {
                    "type": "integer"
                },
                "object_type": {
                    "type": "string"
                },
                "parent_code": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.OperationHistory": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "type": "object"
                },
                "id": {
                    "type": "integer"
                },
                "object_id": {
                    "type": "integer"
                },
                "object_type": {
                    "type": "string"
                },
                "operation_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.ScanEvent": {
            "type": "object",
            "properties": {
                "device_info": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "marker_code": {
                    "type": "string"
                },
                "scanned_at": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.ScanResult": {
            "type": "object",
            "properties": {
                "object": {
                    "$ref": "#/definitions/models.ObjectCard"
                },
                "scan_event": {
                    "$ref": "#/definitions/models.ScanEvent"
                }
            }
        }
    }
}